Revert "Add f32 support"
This reverts commit c80bfc729ff5f08615fcfc1f1802dd915317d9eb.

+++ serde-generate/src/dart.rs
@@ -291,7 +291,7 @@ where
             U32 => "int".into(),
             U64 => "int".into(),
             U128 => "Uint128".into(),
-            F32 => "double".into(),
+            F32 => "float".into(),
             F64 => "double".into(),
             Char => "int".into(),
             Str => "String".into(),
